# [REVIEW] PubSub fast-path subscriber performance - UA_decodeBinary  memory allocation