--- a/examples/pubsub/server_pubsub_subscriber_rt_level.c
+++ b/examples/pubsub/server_pubsub_subscriber_rt_level.c
@@ -67,6 +67,29 @@ externalDataReadNotificationCallback(UA_Server *server, const UA_NodeId *session
     return UA_STATUSCODE_GOOD;
 }
 
+static void
+SubscribeAfterWriteCallback(
+   UA_Server *ipServer,
+   const UA_NodeId *ipDataSetReaderId,
+   const UA_NodeId *ipReaderGroupId,
+   const UA_NodeId *ipTargetVariableId,
+   void *ipTargetVariableContext,
+   UA_DataValue **ippExternalValueSource) { // received value has already been copied to ippExternalValueSource
+
+    (void) ipServer;
+    (void) ipDataSetReaderId;
+    (void) ipReaderGroupId;
+    (void) ipTargetVariableContext;
+
+    UA_String strId;
+    UA_String_init(&strId);
+    UA_NodeId_print(ipTargetVariableId, &strId);
+    UA_LOG_INFO(UA_Log_Stdout, UA_LOGCATEGORY_SERVER, "SubscribeAfterWriteCallback(): "
+        "WriteUpdate() for node Id = '%.*s'. New Value = %u", (UA_Int32) strId.length, strId.data,
+        *((UA_UInt32*) (**ippExternalValueSource).value.data));
+    UA_String_clear(&strId);
+}
+
 /* Define MetaData for TargetVariables */
 static void fillTestDataSetMetaData(UA_DataSetMetaDataType *pMetaData) {
     if(pMetaData == NULL)
@@ -206,6 +229,7 @@ addSubscribedVariables (UA_Server *server) {
         UA_FieldTargetDataType_init(&readerConfig.subscribedDataSet.subscribedDataSetTarget.targetVariables[i].targetVariable);
         readerConfig.subscribedDataSet.subscribedDataSetTarget.targetVariables[i].targetVariable.attributeId  = UA_ATTRIBUTEID_VALUE;
         readerConfig.subscribedDataSet.subscribedDataSetTarget.targetVariables[i].targetVariable.targetNodeId = newnodeId;
+        readerConfig.subscribedDataSet.subscribedDataSetTarget.targetVariables[i].afterWrite = SubscribeAfterWriteCallback;
     }
 
     return retval;