--- a/examples/pubsub_realtime/pubsub_interrupt_publish.c
+++ b/examples/pubsub_realtime/pubsub_interrupt_publish.c
@@ -21,7 +21,6 @@
 #include <open62541/plugin/log_stdout.h>
 #include <open62541/server_pubsub.h>
 #include <open62541/plugin/pubsub_ethernet.h>
-#include "bufmalloc.h"
 
 #define ETH_PUBLISH_ADDRESS      "opc.eth://0a-00-27-00-00-08"
 #define ETH_INTERFACE            "enp0s8"
@@ -106,9 +105,7 @@ publishInterrupt(int sig, siginfo_t* si, void* uc) {
     /* Execute the publish callback in the interrupt */
     struct timespec begin, end;
     clock_gettime(CLOCKID, &begin);
-    useMembufAlloc();
     pubCallback(pubServer, pubData);
-    useNormalAlloc();
     clock_gettime(CLOCKID, &end);
 
     if(publisherMeasurementsCounter >= MAX_MEASUREMENTS)