--- a/examples/pubsub_realtime/vxworks/pubsub_interrupt_publish_tsn.c
+++ b/examples/pubsub_realtime/vxworks/pubsub_interrupt_publish_tsn.c
@@ -22,7 +22,6 @@
 #include <open62541/server_config_default.h>
 #include <open62541/plugin/log_stdout.h>
 #include <open62541/plugin/pubsub_ethernet.h>
-#include "../bufmalloc.h"
 
 #define ETH_PUBLISH_ADDRESS     "opc.eth://01-00-5E-00-00-01"
 #define MILLI_AS_NANO_SECONDS   (1000 * 1000)
@@ -416,7 +415,6 @@ static void open62541EthTSNTask(void) {
             break;
         }
         t = ieee1588TimeGet();
-        useMembufAlloc();
 
         /*
          * Because we can't get the task end time of one packet before it is
@@ -429,7 +427,6 @@ static void open62541EthTSNTask(void) {
             lastTaskEndTime = 0;
         }
         pubCallback(pubServer, pubData);
-        useNormalAlloc();
 
         sequenceNumber++;
         lastCycleTriggerTime = cycleTriggerTime;