Hi!

I have had a look at PubSub Writer and Reader performance on my test hardware.
I am using a custom PubSub  application callback like in your realtime examples with fast-path (ExternalValueCallback) and try to publish/subscribe 256 variables within 1 ms.

I saw that the Publisher needs about 30-40 us and the Reader about 150-160 us to process 256 INT32 variables.
I was wondering why the Reader needed so much more time compared to the Writer.
(My backend variable write / read implementation takes an equal amount of time)

I measured the subscribeCallback steps:

- receive() network message  -> 6 us
- updated buffered network message (decode fields) - 60 us
- write data (memcpy the data to fast-path subscriber variables - ExternalValueCallback pointer) - 8 us
- cleanup (free dataset fields) - 60 us

Decoding the fields and cleanup takes the most time (60 + 60 us)
I thought with fast-path configuration the Reader would directly decode the data to the pointer given by the ExternalValue callback or StaticValueSource.

But this is not the case, because UA_decodeBinary always allocates memory and that's why the cleanup is also necessary.
So in one cycle I have 256 mallocs and frees.

- decodeBinary allocates mem
- then memory is copied to subscriber var
- then the memory is freed

This would not be necessary if UA_decodeBinary would only allocate memory if needed.
Then we could directly decode the data to the given StaticValueSource pointer and save the memory allocation, copying and free.

I have implemented a dirty workaround by checking in the UA_decodeBinary function if the destination is a Variant and if the UA_VARIANT_DATA_NODELETE storageType is set. In this case I do not allocate memory and do not free it and the Reader is now almost as fast as the Writer with about 40 us.

I know that my implementation is not usable, it was just a performance test for me.

But now I don't know how to change the UA_decodeBinary function to only allocate memory if necessary.
Or if I should add a second UA_decodeBinary_noAlloc function or something like this.

Can you give me any suggestions?