@MartinaFelberFEM 2021-03-26
I could write a UA_Variant_decodeBinary_noAlloc() function, which does not allocate memory and is only called at fast-path subscriber UA_NetworkMessage_updateBufferedNwMessage().

And delete the Variant_clear, DataValue_clear calls at subscribeCallback() then I wouldn't have to modify the UA_clear() ...