@jpfr 2021-03-28
Hey Martina,

this is a good push. However, it involves adding hacks to the decoder.
I think this can be done nearly hack-free!

The option UA_ENABLE_MALLOC_SINGLETON switches all uses of malloc/free et al. to global method pointers.
If you are single-threaded, then these can be set at runtime.

There is a simple arena-based malloc here: https://github.com/open62541/open62541/blob/0db99235b08359a3bb5844cfc2d02d36eadd70c0/examples/pubsub_realtime/bufmalloc.c

In the realtime case, you could switch to the arena-based malloc and restore the normal malloc afterwards.
Just make sure not to free memory that now points to the arena.