@MartinaFelberFEM 2021-03-29
Hi!
Many thanks for your response!
I hope I have understood your suggestions.

I would need a multithreading safe implementation.

I have had a look at the bufmalloc.c implementation and I think it can't be used in a multithreading environment with multiple Publishers and Subscribers out of the box. Because there's no locking. And in the examples it's also only used with publisher only (pubsub_interrupt_publish_tsn.c and pubsub_interrupt_publish.c). 

Except I would use the implementation and add a locking mechanism. And it's interesting how long the implementation with the publisher works, because resetMembuf() is never called and there's no free (I think), so sometime the publisher should be out of memory. But you probably could call resetMembuf() and then it would be fine and it's just an example.

But still I would like to avoid the memory allocation (regardless if it's on heap or stack), the copying and the freeing.
Because it's unnecessary. The concept behind ExternalValue should be that the pointer is provided by the application.

And everything is already beautifully prepared (buffer, callbackpointer etc.). It's just not used to it's full potential like in the Writer ...