@jpfr 2021-03-29
Hum. Okay. Multi-threaded.

You could use the singleton malloc pointer as a thread-local variable.
So every thread can switch its malloc individually.

Its a bit more work but looks doable.

1. Add a UA_THREAD_LOCAL define that can be adjusted to architectures and compilers.
https://gcc.gnu.org/onlinedocs/gcc/Thread-Local.html
2. Use UA_THREAD_LOCAL for the singleton malloc pointers.
3. Extend or fork the membuf malloc to avoid races between threads.
4. Switch in and out of custom malloc in the PubSub realtime path.