@MartinaFelberFEM 2021-03-29
Thanks for your help!
I will try that and make time measurements.

But I have changed my hack implementation and added a function UA_Variant_decodeBinary_noAlloc.
And I am working on UA_DataValue_decodeBinary_noAlloc and try to support arrays and structures with no allocation decoding.
Because I would still prefer to not handle any memory operations in the realtime task at all if it can be avoided.

Would these noAlloc functions be ok? Or do you have any other suggestions?
Have there ever been any thoughts about Variant_decoding with preallocated memory?