@MartinaFelberFEM 2021-04-01
I try to test it with the examples server_pubsub_publisher_rt_level and server_pubsub_subscriber_rt_level but the subscriber does not successfully receive messages. -> PubSub receive. Unknown field type.

Setting the dataSetFieldContentMask in subscriber to RAWDATA as is done in the Publisher did not help, only removing the RAWDATA option.