@MartinaFelberFEM 2021-04-01
@jpfr FYI:
Your original idea of switching the thread-local memory like this:

useMemBuf()
updateBufferedNwMessage
useNormalAlloc()
DataSetReaderProcess
-> only clear if build option BUFALLOC is not defined

did not quite work. I need the clear() methods, because they memset the Variant data pointer. If the Variant is not cleared, the UA_Server_unfreezeReaderGroupConfiguration function tries to free it and this causes a segfault.

I am not sure but I think we can't avoid the clear() calls for every field :(
I have tried to adapt the UA_DataSetMessage_clear() function but this does not work.

The memory is allocated at subscribeCallback() and the pointer has to set back to 0 ...