@MartinaFelberFEM 2021-04-01
Regarding the receiveMsgBuffer -> I know you showed me the UA_Byte static thread-local array, but if I use a byte array I would have to change the receive() interface, because it expects a ByteString ...