@jpfr 2021-04-01
You can use _init instead of _clear to reset.