@MartinaFelberFEM 2021-04-02
Thanks, readability is a bit worse with Variant_init() but I hope it's ok

If you are interested: I measured the reader performance
Before optimization: 160 us
EncodeBinary hack: 40 us
MemPool implementation: 60 us

I am very happy about the solution now, many thanks for your help!